introducing sgr.js, export mods, ensure old tests is ok.
