+++ bare.js
@@ -6,37 +6,16 @@ var d              = require('d')
   , map            = require('es5-ext/object/map')
   , primitiveSet   = require('es5-ext/object/primitive-set')
   , setPrototypeOf = require('es5-ext/object/set-prototype-of')
-  , includes       = require('es5-ext/string/#/contains')
   , memoize        = require('memoizee')
   , memoizeMethods = require('memoizee/methods')
+
+  , sgr = require('./lib/sgr')
+  , mods = sgr.mods
 
   , join = Array.prototype.join, defineProperty = Object.defineProperty
   , max = Math.max, min = Math.min
   , variantModes = primitiveSet('_fg', '_bg')
   , xtermMatch, getFn;
-
-var mods = assign({
-	// Style
-	bold:      { _bold: [1, 22] },
-	italic:    { _italic: [3, 23] },
-	underline: { _underline: [4, 24] },
-	blink:     { _blink: [5, 25] },
-	inverse:   { _inverse: [7, 27] },
-	strike:    { _strike: [9, 29] }
-
-	// Color
-}, ['black', 'red', 'green', 'yellow', 'blue', 'magenta', 'cyan', 'white']
-	.reduce(function (obj, color, index) {
-		// foreground
-		obj[color] = { _fg: [30 + index, 39] };
-		obj[color + 'Bright'] = { _fg: [90 + index, 39] };
-
-		// background
-		obj['bg' + color[0].toUpperCase() + color.slice(1)] = { _bg: [40 + index, 49] };
-		obj['bg' + color[0].toUpperCase() + color.slice(1) + 'Bright'] = { _bg: [100 + index, 49] };
-
-		return obj;
-	}, {}));
 
 // Some use cli-color as: console.log(clc.red('Error!'));
 // Which is inefficient as on each call it configures new clc object
@@ -74,12 +53,12 @@ if (process.platform === 'win32') xtermMatch = require('./lib/xterm-match');
 getFn = function () {
 	return setPrototypeOf(function self(/*…msg*/) {
 		var start = '', end = '', msg = join.call(arguments, ' '), conf = self._cliColorData
-		  , hasAnsi = includes.call(msg, '\x1b[');
+		  , hasAnsi = sgr.hasCSI(msg);
 		forEach(conf, function (mod, key) {
-			end = '\x1b[' + mod[1] + 'm' + end;
-			start += '\x1b[' + mod[0] + 'm';
+			end    = sgr(mod[1]) + end;
+			start += sgr(mod[0]);
 			if (hasAnsi) {
-				msg = msg.replace(getEndRe(mod[1]), variantModes[key] ? '\x1b[' + mod[0] + 'm' : '');
+				msg = msg.replace(getEndRe(mod[1]), variantModes[key] ? sgr(mod[0]) : '');
 			}
 		}, null, true);
 		return start + msg + end;
